Fix null pointer exception in excel writer.

+++ engine/org.eclipse.birt.report.engine.emitter.prototype.excel/src/org/eclipse/birt/report/engine/emitter/excel/ExcelWriter.java
@@ -59,10 +59,13 @@ public class ExcelWriter implements IExcelWriter
 	public void end( ) throws IOException
 	{
 		writer.end( );
-		File file = new File( tempFilePath );
-		if ( file.exists( ) && file.isFile( ) )
+		if ( tempFilePath != null )
 		{
-			file.delete( );
+			File file = new File( tempFilePath );
+			if ( file.exists( ) && file.isFile( ) )
+			{
+				file.delete( );
+			}
 		}
 	}
